# api/openapi3.yaml
openapi: 3.0.0
info:
  contact:
    email: support@wishlist.example.com
    name: API Support
  description: >-
    A RESTful API for managing wish lists, gift items, and reservations.

    Features include user authentication, wish list management, gift item tracking, and reservation system.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Wish List API
  version: "1.1"
paths:
...
  /guest/reservations:
    get:
      description: Get all reservations made by a guest using their reservation token.
      parameters:
        - description: Reservation token
          in: query
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of guest reservations retrieved successfully
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/wish-list_internal_domain_reservation_delivery_http\
                    _dto.ReservationDetailsResponse"
                type: array
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      summary: Get reservations made by a guest using a token
      tags:
        - Reservations
...
  /images/upload:
    post:
      description: Upload an image file to S3 storage. The user must be authenticated.
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  description: Image file to upload (max 10MB, only images allowed)
                  type: string
                  format: binary
              required:
                - image
        required: true
      responses:
        "200":
          description: Image uploaded successfully, returns URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/wish-list_internal_domain_storage_delivery_http_dto\
                  .UploadImageResponse"
        "400":
          description: Invalid file or file too large
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Upload an image to S3
      tags:
        - S3 Upload
...
  "/public/reservations/list/{slug}/item/{itemId}":
    get:
      description: Get the reservation status for a specific gift item in a public
        wish list.
      parameters:
        - description: Public wish list slug
          in: path
          name: slug
          required: true
          schema:
            type: string
        - description: Gift Item ID
          in: path
          name: itemId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reservation status retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/wish-list_internal_domain_reservation_delivery_http\
                  _dto.ReservationStatusResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      summary: Get the reservation status for a gift item in a public wish list
      tags:
        - Reservations
...
  /reservations/user:
    get:
      description: Get all reservations made by the authenticated user with pagination.
      parameters:
        - description: Page number (default 1)
          in: query
          name: page
          schema:
            type: integer
        - description: Items per page (default 10, max 100)
          in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: List of user reservations retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/wish-list_internal_domain_reservation_delivery_http\
                  _dto.UserReservationsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Get all reservations made by the authenticated user
      tags:
        - Reservations
  "/reservations/wishlist/{wishlistId}/item/{itemId}":
    delete:
      description: Cancel a reservation for a gift item. Can be done by authenticated
        users or guests (with reservation token).
      parameters:
        - description: Wish List ID
          in: path
          name: wishlistId
          required: true
          schema:
            type: string
        - description: Gift Item ID
          in: path
          name: itemId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/wish-list_internal_domain_reservation_delivery_http\
                _dto.CancelReservationRequest"
        description: Cancellation information (required for guests)
      responses:
        "200":
          description: Reservation canceled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/wish-list_internal_domain_reservation_delivery_http\
                  _dto.CreateReservationResponse"
        "400":
          description: Invalid request body or validation error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "401":
          description: Unauthorized (guests need reservation token)
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Cancel a reservation for a gift item
      tags:
        - Reservations
    post:
      description: Create a reservation for a gift item. Can be done by authenticated
        users or guests (with name and email).
      parameters:
        - description: Wish List ID
          in: path
          name: wishlistId
          required: true
          schema:
            type: string
        - description: Gift Item ID
          in: path
          name: itemId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/wish-list_internal_domain_reservation_delivery_http\
                _dto.CreateReservationRequest"
        description: Reservation information (required for guests)
      responses:
        "200":
          description: Reservation created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/wish-list_internal_domain_reservation_delivery_http\
                  _dto.CreateReservationResponse"
        "400":
          description: Invalid request body or validation error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "401":
          description: Unauthorized (guests need name and email)
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Create a reservation for a gift item
      tags:
        - Reservations
...
servers:
  - url: //localhost:8080/api
components:
  securitySchemes:
    BearerAuth:
      description: Type "Bearer" followed by a space and JWT token.
      in: header
      name: Authorization
      type: apiKey
...